Move allow-local-discovery opt to NetworkConfig
Move the 'allow-local-discovery' option to the NetworkConfig struct (it makes sense to have it there since it's a network configuration option). Also add a member flag to NodeTable which indicates if local discovery
is enabled and move NodeIPEndpoint::isAllowed() to a p2p helper function not bound to the NodeIPEndpoint and update the appropriate tests.